refactor: structural defect resolution — 6 phases
Phase 1: Fix FALLBACK_CROSS_PROVIDER hook wiring (never emitted)
Phase 2: Add get_signal_adapter() ContextVar getter
Phase 3: Fix auto_learn L6→L5 layer violation via DI
Phase 4: Extract CLI memory_handler, scheduler_drain, terminal
Phase 5: Decompose AgenticLoop — context_manager.py, convergence.py
Phase 6: Split Runtime.create() into 4 staged builders

Metrics: agentic_loop 1818→1405 (-23%), cli/__init__ 1892→1641 (-13%)
Quality: ruff 0, mypy 0, pytest 3702 pass
