+++ core/agent/agentic_loop.py
@@ -250,12 +250,18 @@ class AgenticLoop:
         self._ESCALATION_THRESHOLD: int = 2
         self._LLM_RETRY_CAP: int = 5  # max retries before giving up
 
-        # Feature 5: Backpressure on consecutive tool failures
-        self._total_consecutive_tool_errors: int = 0
+        # Context window management (extracted — SRP)
+        from core.agent.context_manager import ContextWindowManager
 
-        # Feature 6: Convergence detection (stuck loop)
-        self._recent_errors: list[str] = []
-        self._convergence_escalated: bool = False
+        self._ctx_mgr = ContextWindowManager(hooks=hooks, quiet=quiet)
+
+        # Convergence detection + tool error tracking (extracted — SRP)
+        from core.agent.convergence import ConvergenceDetector
+
+        # Late-binding lambda so test patches on _try_model_escalation propagate
+        self._convergence = ConvergenceDetector(
+            escalation_fn=lambda: self._try_model_escalation()
+        )
 
         # Feature 8: Diversity forcing — prevent same tool 5x consecutively
         self._consecutive_tool_tracker: list[str] = []
@@ -968,10 +974,12 @@ class AgenticLoop:
             if self._check_convergence_break():
                 from core.cli.ui.agentic_ui import emit_convergence_detected
 
-                emit_convergence_detected(
-                    self._recent_errors[-1] if self._recent_errors else "unknown",
-                    round_idx + 1,
+                last_err = (
+                    self._convergence.recent_errors[-1]
+                    if self._convergence.recent_errors
+                    else "unknown"
                 )
+                emit_convergence_detected(last_err, round_idx + 1)
                 self._op_logger.finalize()
                 self._sync_messages_to_context(messages)
                 result = AgenticResult(
@@ -985,11 +993,11 @@ class AgenticLoop:
                 )
                 return self._finalize_and_return(result, user_input, round_idx + 1)
 
-            if self._total_consecutive_tool_errors >= 3:
+            if self._convergence.total_consecutive_tool_errors >= 3:
                 # Backpressure: inject a cooldown hint
                 from core.cli.ui.agentic_ui import emit_tool_backpressure
 
-                emit_tool_backpressure(self._total_consecutive_tool_errors)
+                emit_tool_backpressure(self._convergence.total_consecutive_tool_errors)
                 await asyncio.sleep(1.0)
                 backpressure_hint = {
                     "type": "text",
@@ -1093,390 +1101,32 @@ class AgenticLoop:
         """
         self.context.messages = list(messages)
 
+    def _notify_context_event(
+        self, event_type: str, *, original_count: int, new_count: int
+    ) -> None:
+        """Notify user of context compression. Delegates to ContextWindowManager."""
+        self._ctx_mgr._notify_context_event(
+            event_type, original_count=original_count, new_count=new_count
+        )
+
     def _maybe_prune_messages(self, messages: list[dict[str, Any]]) -> None:
-        """Prune old messages when conversation exceeds 5 rounds (10 msgs).
-
-        Keeps first user message + bridge + recent messages for context budget.
-        Ensures:
-        1. user/assistant alternation is preserved
-        2. No orphaned tool_result messages (each tool_result must follow
-           an assistant message containing the matching tool_use block)
-        """
-        if len(messages) <= 30:
-            return
-        first = messages[0]
-        # Walk backward to find a safe cut point:
-        # - Must be a "user" role message
-        # - Must NOT be a tool_result message (those need a preceding tool_use)
-        # Start from -4 and go back until we find a plain user text message
-        safe_cut = None
-        for candidate in range(-4, -(len(messages)), -1):
-            idx = len(messages) + candidate
-            if idx <= 0:
-                break
-            msg = messages[idx]
-            if msg["role"] != "user":
-                continue
-            # Check if this is a tool_result message
-            content = msg.get("content")
-            if isinstance(content, list) and any(
-                isinstance(b, dict) and b.get("type") == "tool_result" for b in content
-            ):
-                continue  # Skip — orphaned tool_result after pruning
-            safe_cut = candidate
-            break
-
-        if safe_cut is None:
-            # No safe cut found — don't prune to avoid corruption
-            return
-
-        recent = messages[safe_cut:]
-        bridge: dict[str, Any] = {
-            "role": "assistant",
-            "content": [{"type": "text", "text": "(earlier rounds omitted)"}],
-        }
-        messages.clear()
-        messages.extend([first, bridge, *recent])
-        log.debug("Pruned messages: kept first + bridge + %d recent", len(recent))
+        """Prune old messages. Delegates to ContextWindowManager."""
+        self._ctx_mgr.maybe_prune_messages(messages)
 
     def _check_context_overflow(self, system: str, messages: list[dict[str, Any]]) -> None:
-        """Check context window usage and apply provider-aware compression.
-
-        Strategy by provider:
-        - Anthropic: server-side compaction (compact_20260112) handles 80%+ automatically.
-          Client only intervenes at 95% as emergency prune safety net.
-        - OpenAI/GLM: no server-side compaction. Client triggers LLM-based compaction
-          at 80% and emergency prune at 95%.
-
-        Absolute ceiling (200K tokens):
-        - Large-context models (1M) hit rate limit pool separation at >200K tokens.
-          Percentage thresholds (80%=800K) are too distant to catch this.
-          When estimated > 200K on a >200K-window model, force tool result
-          summarization + compact to stay under the ceiling.
-
-        Compression strategy is delegated to the CONTEXT_OVERFLOW_ACTION hook handler.
-        If no handler is registered or all fail, falls back to hardcoded defaults.
-        """
-        try:
-            from core.config import settings
-            from core.orchestration.context_monitor import check_context
-
-            metrics = check_context(messages, self.model, system_prompt=system)
-
-            if metrics.is_critical:
-                log.warning(
-                    "Context CRITICAL: %.0f%% (%d/%d tokens) — emergency action",
-                    metrics.usage_pct,
-                    metrics.estimated_tokens,
-                    metrics.context_window,
-                )
-                if self._hooks:
-                    self._hooks.trigger(
-                        HookEvent.CONTEXT_CRITICAL,
-                        {"metrics": dataclasses.asdict(metrics), "model": self.model},
-                    )
-
-                strategy = self._resolve_overflow_strategy(metrics, settings)
-                self._apply_overflow_strategy(strategy, messages, settings)
-
-                # Re-check: if still critical after pruning, context is exhausted
-                post = check_context(messages, self.model, system_prompt=system)
-                if post.is_critical:
-                    raise _ContextExhaustedError(
-                        f"Context exhausted: {post.usage_pct:.0f}% after pruning"
-                    )
-
-            elif metrics.is_warning:
-                # Step 1: mask stale observations (cheapest — no LLM call)
-                from core.orchestration.context_monitor import mask_stale_observations
-
-                mask_keep = settings.observation_mask_keep_rounds
-                masked = mask_stale_observations(messages, keep_recent_rounds=mask_keep)
-                if masked > 0:
-                    log.info(
-                        "Context at %.0f%%: masked %d stale observations",
-                        metrics.usage_pct,
-                        masked,
-                    )
-
-                # Step 2: compact or summarize
-                strategy = self._resolve_overflow_strategy(metrics, settings)
-                if strategy.get("strategy") == "compact":
-                    self._apply_overflow_strategy(strategy, messages, settings)
-                else:
-                    from core.orchestration.context_monitor import summarize_tool_results
-
-                    summarized, _tok_before, _tok_after = summarize_tool_results(
-                        messages,
-                        metrics.context_window,
-                    )
-                    if summarized > 0:
-                        log.info(
-                            "Context at %.0f%%: summarized %d large tool results",
-                            metrics.usage_pct,
-                            summarized,
-                        )
-
-            elif metrics.is_ceiling_exceeded:
-                # Absolute 200K ceiling breached on a large-context model.
-                # Percentage is low (e.g. 20-30%) but tokens exceed the rate limit
-                # pool boundary. Summarize tool results first (lightest), then
-                # compact if still over.
-                from core.orchestration.context_monitor import (
-                    ABSOLUTE_TOKEN_CEILING,
-                    summarize_tool_results,
-                )
-
-                log.info(
-                    "Context ceiling: %d tokens > %dK ceiling (%.0f%% of %dK window) "
-                    "— compressing to avoid rate limit pool separation",
-                    metrics.estimated_tokens,
-                    ABSOLUTE_TOKEN_CEILING // 1000,
-                    metrics.usage_pct,
-                    metrics.context_window // 1000,
-                )
-
-                # Phase 1: summarize large tool results
-                summarized, _tok_before, _tok_after = summarize_tool_results(
-                    messages,
-                    ABSOLUTE_TOKEN_CEILING,
-                )
-                post = check_context(messages, self.model, system_prompt=system)
-
-                if post.is_ceiling_exceeded:
-                    # Phase 2: compact conversation
-                    strategy = {"strategy": "compact", "keep_recent": settings.compact_keep_recent}
-                    self._apply_overflow_strategy(strategy, messages, settings)
-
-        except Exception:
-            log.debug("Context monitor check failed", exc_info=True)
-
-    def _apply_overflow_strategy(
-        self,
-        strategy: dict[str, Any],
-        messages: list[dict[str, Any]],
-        settings: Any,
-    ) -> None:
-        """Execute the overflow strategy (prune or compact)."""
-        from core.orchestration.context_monitor import prune_oldest_messages
-
-        action = strategy.get("strategy", "none")
-        keep_recent = strategy.get("keep_recent", settings.compact_keep_recent)
-
-        if action == "compact":
-            import asyncio
-
-            from core.orchestration.compaction import compact_conversation
-
-            try:
-                new_msgs, did_compact = asyncio.get_event_loop().run_until_complete(
-                    compact_conversation(
-                        messages,
-                        provider=self._provider,
-                        model=self.model,
-                        keep_recent=keep_recent,
-                    )
-                )
-                if did_compact:
-                    original_count = len(messages)
-                    messages.clear()
-                    messages.extend(new_msgs)
-                    self._notify_context_event(
-                        "compact",
-                        original_count=original_count,
-                        new_count=len(new_msgs),
-                    )
-                    return
-            except Exception:
-                log.warning("Client compaction failed — falling back to prune", exc_info=True)
-            # Fall through to prune on failure
-            action = "prune"
-
-        if action == "prune":
-            pruned = prune_oldest_messages(messages, keep_recent=keep_recent)
-            original_count = len(messages)
-            if len(pruned) < original_count:
-                messages.clear()
-                messages.extend(pruned)
-                log.info(
-                    "Emergency pruned: %d → %d messages (keep_recent=%d)",
-                    original_count,
-                    len(pruned),
-                    keep_recent,
-                )
-                self._notify_context_event(
-                    "prune",
-                    original_count=original_count,
-                    new_count=len(pruned),
-                )
+        """Check context window usage. Delegates to ContextWindowManager."""
+        self._ctx_mgr.check_context_overflow(system, messages, self.model, self._provider)
 
     def _aggressive_context_recovery(self, system: str, messages: list[dict[str, Any]]) -> int:
-        """Last-resort context recovery: aggressive prune + tool result summarization.
-
-        Called when normal compaction/prune at 95% still leaves context critical.
-        Two-phase approach:
-          1. Summarize large tool results (>2% of window) to compact summaries
-          2. Prune with halved keep_recent (keep fewer messages)
-        Returns number of messages freed, or 0 if recovery failed.
-        """
-        try:
-            from core.config import settings
-            from core.orchestration.context_monitor import (
-                check_context,
-                prune_oldest_messages,
-                summarize_tool_results,
-            )
-
-            original_count = len(messages)
-
-            # Phase 1: summarize large tool_result blocks in-place
-            ctx_window = getattr(
-                check_context(messages, self.model, system_prompt=system),
-                "context_window",
-                200_000,
-            )
-            summarized, _tok_before, _tok_after = summarize_tool_results(messages, ctx_window)
-            if summarized > 0:
-                log.info("Aggressive recovery: summarized %d tool results", summarized)
-
-            # Check if summarization alone resolved it
-            post = check_context(messages, self.model, system_prompt=system)
-            if not post.is_critical:
-                return original_count - len(messages) + summarized
-
-            # Phase 2: prune with halved keep_recent
-            aggressive_keep = max(3, settings.compact_keep_recent // 2)
-            pruned = prune_oldest_messages(messages, keep_recent=aggressive_keep)
-            if len(pruned) < len(messages):
-                messages.clear()
-                messages.extend(pruned)
-                log.info(
-                    "Aggressive prune: %d → %d messages (keep_recent=%d)",
-                    original_count,
-                    len(pruned),
-                    aggressive_keep,
-                )
-
-            # Final check
-            post2 = check_context(messages, self.model, system_prompt=system)
-            if not post2.is_critical:
-                return original_count - len(messages)
-
-            return 0  # recovery failed
-        except Exception:
-            log.warning("Aggressive context recovery failed", exc_info=True)
-            return 0
-
-    def _notify_context_event(
-        self,
-        event_type: str,
-        *,
-        original_count: int,
-        new_count: int,
-    ) -> None:
-        """Notify user of automatic context compression via UI."""
-        if self._quiet:
-            return
-        try:
-            from core.cli.ui.agentic_ui import render_context_event
-
-            render_context_event(event_type, original_count=original_count, new_count=new_count)
-        except Exception:
-            log.debug("Context event notification failed", exc_info=True)
-
-    def _resolve_overflow_strategy(self, metrics: Any, settings: Any) -> dict[str, Any]:
-        """Ask CONTEXT_OVERFLOW_ACTION hook for compression strategy, with fallback.
-
-        Returns a dict with at least ``strategy`` key. If no handler responds
-        or all handlers fail, returns the hardcoded default strategy.
-
-        Provider-aware: passes self._provider so the hook can differentiate
-        between Anthropic (server-side compaction) and others (client-side).
-        """
-        if self._hooks:
-            results = self._hooks.trigger_with_result(
-                HookEvent.CONTEXT_OVERFLOW_ACTION,
-                {
-                    "metrics": dataclasses.asdict(metrics),
-                    "model": self.model,
-                    "provider": self._provider,
-                },
-            )
-            for result in results:
-                if result.success and result.data.get("strategy"):
-                    return result.data
-
-        # Fallback: hardcoded default (no handler registered or all failed)
-        keep_recent = settings.compact_keep_recent
-        if self._provider == "anthropic":
-            # Anthropic: server-side handles it, only emergency prune at 95%
-            if metrics.usage_pct >= 95:
-                return {"strategy": "prune", "keep_recent": keep_recent}
-            return {"strategy": "none"}
-
-        # Non-Anthropic: compact at 80%, prune at 95%
-        if metrics.context_window < 200_000:
-            keep_recent = min(keep_recent, 5)
-        if metrics.usage_pct >= 95:
-            return {"strategy": "prune", "keep_recent": keep_recent}
-        elif metrics.usage_pct >= 80:
-            return {"strategy": "compact", "keep_recent": keep_recent}
-        return {"strategy": "none"}
+        """Last-resort context recovery. Delegates to ContextWindowManager."""
+        return self._ctx_mgr.aggressive_context_recovery(system, messages, self.model)
 
     @staticmethod
     def _repair_messages(messages: list[dict[str, Any]]) -> None:
-        """Remove orphaned tool_result messages that lack a preceding tool_use.
+        """Remove orphaned tool_result messages. Delegates to ContextWindowManager."""
+        from core.agent.context_manager import ContextWindowManager
 
-        Scans backward and removes any user message whose content is entirely
-        tool_result blocks without matching tool_use in the prior assistant msg.
-        Also syncs the repair back to context via mutation in place.
-        """
-        i = len(messages) - 1
-        while i >= 1:
-            msg = messages[i]
-            if msg["role"] != "user":
-                i -= 1
-                continue
-            content = msg.get("content")
-            if not isinstance(content, list):
-                i -= 1
-                continue
-            # Check if ALL blocks are tool_result
-            tr_ids = {
-                b["tool_use_id"]
-                for b in content
-                if isinstance(b, dict) and b.get("type") == "tool_result"
-            }
-            if not tr_ids:
-                i -= 1
-                continue
-            # Check preceding assistant message for matching tool_use
-            if i > 0 and messages[i - 1]["role"] == "assistant":
-                prev_content = messages[i - 1].get("content", [])
-                if isinstance(prev_content, list):
-                    tu_ids = {
-                        b.get("id")
-                        for b in prev_content
-                        if isinstance(b, dict) and b.get("type") == "tool_use"
-                    }
-                    if tr_ids <= tu_ids:
-                        i -= 1
-                        continue  # All tool_results have matching tool_use — OK
-            # Orphaned — remove this tool_result message and its preceding
-            # assistant message (which also lost its tool_use context)
-            log.debug("Removing orphaned tool_result at index %d", i)
-            messages.pop(i)
-            # Also remove the preceding assistant message if it's a bridge/text
-            if i > 0 and messages[i - 1]["role"] == "assistant":
-                prev_c = messages[i - 1].get("content", [])
-                if isinstance(prev_c, list):
-                    has_tool_use = any(
-                        isinstance(b, dict) and b.get("type") == "tool_use" for b in prev_c
-                    )
-                    if not has_tool_use:
-                        messages.pop(i - 1)
-                        i -= 1
-            i -= 1
+        ContextWindowManager.repair_messages(messages)
 
     def _build_system_prompt(self) -> str:
         """Build the system prompt with skill context and agentic suffix."""
@@ -1707,19 +1357,35 @@ class AgenticLoop:
                 return True
 
         # Current provider's chain exhausted — try cross-provider (Feature 4)
-        fallbacks = CROSS_PROVIDER_FALLBACK.get(self._provider, [])
+        from_provider = self._provider
+        fallbacks = CROSS_PROVIDER_FALLBACK.get(from_provider, [])
         for fallback_provider, fallback_model in fallbacks:
             if fallback_model != current:
                 log.warning(
                     "Cross-provider escalation: %s(%s) -> %s(%s)",
                     current,
-                    self._provider,
+                    from_provider,
                     fallback_model,
                     fallback_provider,
                 )
                 self.update_model(
                     fallback_model, fallback_provider, reason="cross_provider_escalation"
                 )
+                # Emit dedicated cross-provider hook for observability
+                # (MODEL_SWITCHED is also fired by update_model, but this
+                # event carries provider-level context for audit loggers)
+                if self._hooks:
+                    from core.hooks import HookEvent
+
+                    self._hooks.trigger(
+                        HookEvent.FALLBACK_CROSS_PROVIDER,
+                        {
+                            "from_model": current,
+                            "to_model": fallback_model,
+                            "from_provider": from_provider,
+                            "to_provider": fallback_provider,
+                        },
+                    )
                 return True
 
         log.warning("Model escalation failed: no more fallback models available")
@@ -1731,88 +1397,11 @@ class AgenticLoop:
     # ---------------------------------------------------------------------------
 
     def _update_tool_error_tracking(self, tool_results: list[dict[str, Any]]) -> None:
-        """Update consecutive tool error tracking and recent error history.
-
-        Processes a batch of tool results from the current round.
-        Resets the consecutive counter on any success, increments on all-error rounds.
-        Appends normalized error keys to _recent_errors for convergence detection.
-        """
-        has_success = False
-        has_error = False
-
-        for tr in tool_results:
-            # Parse content JSON to check for errors
-            content = tr.get("content", "")
-            if isinstance(content, str):
-                try:
-                    parsed = json.loads(content)
-                except (json.JSONDecodeError, ValueError):
-                    parsed = {}
-            else:
-                parsed = content if isinstance(content, dict) else {}
-
-            if isinstance(parsed, dict) and parsed.get("error"):
-                has_error = True
-                # Feature 6: append normalized error key
-                tool_use_id = tr.get("tool_use_id", "")
-                error_str = str(parsed.get("error", ""))[:50]
-                # Extract tool name from tool_log (most recent entries)
-                tool_name = "unknown"
-                for entry in reversed(self._tool_processor.tool_log):
-                    if tool_use_id and entry.get("tool") and isinstance(entry.get("result"), dict):
-                        tool_name = entry["tool"]
-                        break
-                error_key = f"{tool_name}:{error_str}"
-                self._recent_errors.append(error_key)
-                # Keep last 6 entries max
-                if len(self._recent_errors) > 6:
-                    self._recent_errors = self._recent_errors[-6:]
-            else:
-                has_success = True
-
-        if has_success:
-            self._total_consecutive_tool_errors = 0
-        elif has_error:
-            self._total_consecutive_tool_errors += 1
+        """Update tool error tracking. Delegates to ConvergenceDetector."""
+        self._convergence.update_tool_error_tracking(
+            tool_results, self._tool_processor.tool_log
+        )
 
     def _check_convergence_break(self) -> bool:
-        """Check if the loop is stuck in a repeating failure pattern.
-
-        On first detection of 3 identical errors, attempts model escalation
-        (runtime ratchet — Karpathy P4) instead of breaking immediately.
-        Only breaks after escalation has been tried and errors persist.
-        """
-        if len(self._recent_errors) < 3:
-            return False
-
-        # Check last 3 entries for identical pattern
-        last_3 = self._recent_errors[-3:]
-        if last_3[0] == last_3[1] == last_3[2]:
-            # Runtime ratchet: try model escalation before giving up
-            if not self._convergence_escalated:
-                self._convergence_escalated = True
-                log.warning(
-                    "Convergence detected (%s x3) — escalating model",
-                    last_3[0],
-                )
-                escalated = self._try_model_escalation()
-                if escalated:
-                    self._recent_errors.clear()
-                    return False  # Give escalated model a chance
-                # Escalation failed (no fallback) — fall through to break check
-
-            # Already escalated and still stuck — check for 4+ identical
-            if len(self._recent_errors) >= 4:
-                last_4 = self._recent_errors[-4:]
-                if last_4[0] == last_4[1] == last_4[2] == last_4[3]:
-                    log.warning(
-                        "Convergence detected after escalation: 4+ identical errors '%s'",
-                        last_4[0],
-                    )
-                    return True
-            # 3 identical post-escalation — log warning, don't break yet
-            log.warning(
-                "Convergence warning (post-escalation): 3 identical errors '%s'",
-                last_3[0],
-            )
-        return False
+        """Check for stuck loop. Delegates to ConvergenceDetector."""
+        return self._convergence.check_convergence_break()
