+++ core/hooks/auto_learn.py
@@ -138,8 +138,15 @@ def detect_patterns(
 # ---------------------------------------------------------------------------
 
 
-def make_auto_learn_handler() -> tuple[str, Callable[..., None]]:
+def make_auto_learn_handler(
+    profile_provider: Callable[[], Any] | None = None,
+) -> tuple[str, Callable[..., None]]:
     """Create a TURN_COMPLETE handler that auto-learns user patterns.
+
+    Args:
+        profile_provider: Callable returning the current user profile.
+            Injected at registration to avoid L6→L5 layer violation.
+            Falls back to lazy import if not provided (backwards compat).
 
     Returns (handler_name, handler_fn) for hook registration.
     """
@@ -148,15 +155,21 @@ def make_auto_learn_handler() -> tuple[str, Callable[..., None]]:
     last_learn_ts = 0.0
     tool_counter: Counter[str] = Counter()
 
+    def _get_profile() -> Any:
+        if profile_provider is not None:
+            return profile_provider()
+        # Fallback for backwards compat (tests that don't inject)
+        from core.tools.profile_tools import get_user_profile
+
+        return get_user_profile()
+
     def _on_turn_complete(event: HookEvent, data: dict[str, Any]) -> None:
         nonlocal session_count, last_learn_ts
 
         if session_count >= _MAX_PER_SESSION:
             return
 
-        from core.tools.profile_tools import get_user_profile
-
-        profile = get_user_profile()
+        profile = _get_profile()
         if profile is None:
             return
